requirements.txt
gunicorn
flask
numpy
matplotlib
pillow
tensorflow-cpu